🔀 Refactor filter components and add store provider
- Refactored the `AllergensFilter` component to use the `useStore` hook from the `StoreProvider`
- Removed unused import statements in both `AllergensFilter` and `DietFilter`
- Moved the state for selected diet and allergens to the `StoreProvider`
- Created a new function, `toggleAllergen`, in the `StoreProvider` to handle toggling of allergens
- Updated references to selected diet and allergens in both filter components

These changes improve code organization and make it easier to manage state related to filters.
