+++ src/providers/StoreProvider.js
@@ -5,13 +5,33 @@ const StoreContext = createContext()
 
 export function StoreProvider({ children }) {
 	const [isFilterModalClosed, setIsFilterModalClosed] = useState(true)
+	const [selectedDiet, setSelectedDiet] = useState('default')
+	const [selectedAllergens, setSelectedAllergens] = useState([])
+
+	const toggleAllergen = key => {
+		if (selectedAllergens.includes(key)) {
+			setSelectedAllergens(prev => prev.filter(item => item !== key))
+		} else {
+			setSelectedAllergens(prev => [...prev, key])
+		}
+	}
 
 	const toggleFilterModal = () => {
 		setIsFilterModalClosed(!isFilterModalClosed)
 	}
 
 	return (
-		<StoreContext.Provider value={{ isFilterModalClosed, toggleFilterModal }}>
+		<StoreContext.Provider
+			value={{
+				isFilterModalClosed,
+				toggleFilterModal,
+				selectedDiet,
+				selectedAllergens,
+				toggleAllergen,
+				setSelectedDiet,
+				setSelectedAllergens,
+			}}
+		>
 			{children}
 		</StoreContext.Provider>
 	)
